Use polling instead of chars_available_callback
When using USB serial, sending large amounts of text back over the port
in the chars_available_callback causes bytes to be lost.

This is kind of a cheap solution - calling getchar_timeout_us(0) in the
main loop - but it works.  When the USB serial port is connected, the
loop executes faster to make it more responsive.